# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management System
  description: |-
    This is a simple REST API for task management system.
    
    Some useful links:
    
    - [The GitHub repository](https://github.com/VVSenti/TaskManagementSystem)
    - [The source API definition](https://github.com/VVSenti/TaskManagementSystem/blob/master/src/main/resources/openapi.yaml)
    
  contact:
    email: sentyurinvv@gmail.com
  version: 1.0.0
...
paths:
...
  /tasks:
    get:
      tags:
        - task
      summary: Get all tasks
      description: Returnes all tasks
      operationId: getAllTasks
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: The request is unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/responses/unauthorized'
      security:
        - api_key: []
    post:
      tags:
        - task
      summary: Create new task
      description: Creation of a new task
      operationId: createTask
      requestBody:
        description: Infrormaion about task, executor email is optional 
        content:
          application/json:
            schema:
              $ref: '#/components/requestBodies/Task'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                example: "OK"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/responses/badRequest'
        '401':
          description: The request is unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/responses/unauthorized'
      security:
        - api_key: []
  /tasks/{taskId}:
    get:
      tags:
        - task
      summary: Task info
      description: Returnes information about the task
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          description: ID of task to return
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: The request is unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/responses/unauthorized'
        '404':
          description: Task hasn't been found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/badRequest'
      security:
        - api_key: []
    delete:
      tags:
        - task
      summary: Delete task
      description: Deletes the task
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          description: ID of task to return
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                example: "OK"
        '401':
          description: The request is unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/responses/unauthorized'
        '404':
          description: Task hasn't been found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/badRequest'
      security:
        - api_key: []
    patch:
      tags:
        - task
      summary: Edit task
      description: Edit information about task
      operationId: editTask
      requestBody:
        description: Information about the task that should be edited. You can leave out fields that will not be changed.
        content:
          application/json:
            schema:
              $ref: '#/components/requestBodies/Task'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                example: "OK"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/responses/badRequest'
        '401':
          description: The request is unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/responses/unauthorized'
        '403':
          description: Only author and executor may edit the task
          content:
            application/json:
              schema:
                $ref: '#/components/responses/unauthorized'
        '404':
          description: Task hasn't been found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/badRequest'
      security:
        - api_key: []
  /tasks/author:
    get:
      tags:
        - task
      summary: Get all tasks created by the author
      description: Returnes all tasks created by the author with specified id
      operationId: getAllTasksOfAuthor
      parameters:
        - name: id
          in: header
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: The request is unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/responses/unauthorized'
        '404':
          description: User with such id hasn't been found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/badRequest'
      security:
        - api_key: []
  /tasks/executor:
    get:
      tags:
        - task
      summary: Get all tasks executing by the user
      description: Get all tasks executing by the user with specified id
      operationId: getAllTasksOfExecutor
      parameters:
        - name: id
          in: header
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: The request is unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/responses/unauthorized'
        '404':
          description: User with such id hasn't been found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/badRequest'
      security:
        - api_key: []
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        title:
          type: string
          example: API documentation
        description:
          type: string
          example: Use Open API!
        status:
          type: string
          description: Task status
          enum:
            - Running
            - Waiting
            - Done
        priority:
          type: string
          description: Task priority
          enum:
            - High
            - Medium
            - Low
        authorEmail:
          type: string
          description: E-mail of a user how are doing this task
          example: kakapo@gmail.com
        executorEmail:
          type: string
          description: E-mail of a user how are doing this task
          example: parrot@gmail.com
...
  requestBodies:
    Task:
      type: object
      properties:
        title:
          type: string
          example: API documentation
        description:
          type: string
          example: Use Open API!
        status:
          type: string
          description: Task status
          enum:
            - Running
            - Waiting
            - Done
        priority:
          type: string
          description: Task priority
          enum:
            - High
            - Medium
            - Low
        executorEmail:
          type: string
          description: E-mail of a user how are doing this task
          exampl: parrot@gmail.com
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header
  responses:
...
    badRequest:
      type: object
      description: Incorrect request body
      properties:
        message:
          type: string
          example: Error description
        timestamp:
          type: string
          example: 2024-03-08T20:00:53.097+00:00
    unauthorized:
      type: object
      description: Unauthorizes request
      properties:
        timestamp:
          type: string
          example: 2024-03-08T20:03:40.190+00:00
        status:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        path:
          type: string
          example: /path